Refactor SecurityConfig and enhance authentication setup.
Renamed SecurityConfig package and added dependency injection for UserDetailsService. Introduced DaoAuthenticationProvider and AuthenticationManager beans to support custom authentication logic with BCryptPasswordEncoder.
